requirements.txt
fastapi
uvicorn[standard]
openai
Pillow
pydantic
python-dotenv
requests
google-genai
passlib[bcrypt]
python-jose[cryptography]
python-multipart
starlette
bcrypt==4.0.1